+++ core/src/main/kotlin/model/transaction/TransactionShare.kt
@@ -4,4 +4,4 @@ import model.User
 
 
 class TransactionShare(val user: User,
-                       val amount: Int)
+                       val amount: Long)
