+++ core/src/main/kotlin/storage/dao/TransactionDao.kt
@@ -9,7 +9,7 @@ import org.jetbrains.exposed.sql.Table
 
 object Transactions : LongIdTable() {
     val fund = reference("fund", Funds).index()
-    val amount = integer("amount")
+    val amount = long("amount")
     val description = varchar("description", 1000)
     val timestamp = datetime("timestamp")
     val status = enumeration("status", TransactionStatus::class.java)
@@ -18,7 +18,7 @@ object Transactions : LongIdTable() {
 object TransactionShares : Table() {
     val transaction = reference("transaction", Transactions).index()
     val user = reference("user", Users).index()
-    val amount = integer("amount")
+    val amount = long("amount")
 
     init {
         this.index(true, transaction, user)
